Rename xcompletion into xis_complete
