+++ src/xinterpreter.cpp
@@ -20,12 +20,12 @@
 
 #include "xeus-python/xinterpreter.hpp"
 #include "xeus-python/xeus_python_config.hpp"
-#include "xcompletion.hpp"
-#include "xinput.hpp"
 #include "xcomm.hpp"
-#include "xinspect.hpp"
-#include "xstream.hpp"
 #include "xdisplay.hpp"
+#include "xinput.hpp"
+#include "xinspect.hpp"
+#include "xis_complete.hpp"
+#include "xstream.hpp"
 #include "xtraceback.hpp"
 #include "xutils.hpp"
